LayoutTests/platform/wk2/http/tests/security/mixedContent/secure-redirect-to-insecure-redirect-to-basic-auth-secure-image-allowCrossOriginSubresourcesToAskForCredentials.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/secure-redirect-to-insecure-redirect-to-basic-auth-secure-image-allowCrossOriginSubresourcesToAskForCredentials.https.html was allowed to display insecure content from http://127.0.0.1:8080/resources/redirect.php?url=https://localhost:8443/security/mixedContent/resources/subresource/protected-image.php.

localhost:8443 - didReceiveAuthenticationChallenge - Responding with testUser:testPassword
This test loads a secure image that redirects to an insecure image that redirects to a secure image guarded by basic authentication. The image should load.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS did load image.
PASS successfullyParsed is true

TEST COMPLETE
